requirements.txt
streamlit
PyMuPDF
Pillow
requests
reportlab
google-genai
markdown
fpdf2